Expand lap time API documentation
