+++ lap_time_api/requirements.txt
@@ -1,0 +1,2 @@
+fastapi
+uvicorn[standard]
