+++ webapp/main.py
@@ -27,13 +27,6 @@ def get_historical_results(season: int, round_: int) -> list[dict]:
     return results.to_dict(orient="records")
 
 
-def get_practice_laps(season: int, round_: int, session: int) -> list[dict]:
-    """Load lap times for a practice session."""
-    ses = fastf1.get_session(season, round_, f"FP{session}")
-    ses.load()
-    laps = ses.laps.fillna("").astype(str)
-    return laps.to_dict(orient="records")
-
 @app.get("/")
 async def index(request: Request):
     return templates.TemplateResponse("index.html", {"request": request})
@@ -44,10 +37,6 @@ async def historical(season: int, round: int):
     return jsonable_encoder(data)
 
 
-@app.get("/practice-laps")
-async def practice_laps(season: int, round: int, session: int = 1):
-    data = get_practice_laps(season, round, session)
-    return jsonable_encoder(data)
 
 
 @app.post("/admin/start")
